Refactor new_network -> network, replacing old network component
