--- packages/client-api/src/user-config/window/providers/new-network-provider-config.model.ts
@@ -1,13 +1,0 @@
-import { z } from 'zod';
-
-import { ProviderType } from '../provider-type.model';
-
-export const NewNetworkProviderConfigSchema = z.object({
-  type: z.literal(ProviderType.NEWNETWORK),
-
-  refresh_interval: z.coerce.number().default(5 * 1000),
-});
-
-export type NewNetworkProviderConfig = z.infer<
-  typeof NewNetworkProviderConfigSchema
->;
