+++ packages/desktop/src/providers/config.rs
@@ -6,7 +6,7 @@ use super::{
   battery::BatteryProviderConfig, cpu::CpuProviderConfig,
   host::HostProviderConfig, ip::IpProviderConfig,
   memory::MemoryProviderConfig, network::NetworkProviderConfig,
-  newnetwork::NewNetworkProviderConfig, weather::WeatherProviderConfig,
+  weather::WeatherProviderConfig,
 };
 
 #[derive(Deserialize, Debug)]
@@ -20,6 +20,5 @@ pub enum ProviderConfig {
   Komorebi(KomorebiProviderConfig),
   Memory(MemoryProviderConfig),
   Network(NetworkProviderConfig),
-  NewNetwork(NewNetworkProviderConfig),
   Weather(WeatherProviderConfig),
 }
