+++ packages/desktop/src/providers/network/config.rs
@@ -3,7 +3,7 @@ use serde::Deserialize;
 use crate::impl_interval_config;
 
 #[derive(Deserialize, Debug)]
-#[serde(tag = "type", rename = "network")]
+#[serde(tag = "type", rename = "newnetwork")]
 pub struct NetworkProviderConfig {
   pub refresh_interval: u64,
 }
