+++ packages/desktop/src/providers/network/mod.rs
@@ -1,6 +1,7 @@
 mod config;
 mod provider;
 mod variables;
+mod parse_netsh;
 
 pub use config::*;
 pub use provider::*;
