+++ packages/desktop/src/providers/network/parse_netsh.rs
