+++ packages/desktop/src/providers/network/provider.rs
@@ -2,21 +2,30 @@ use std::sync::Arc;
 
 use anyhow::Result;
 use async_trait::async_trait;
-use sysinfo::Networks;
+use netdev::interface::{
+  get_default_interface, get_interfaces, Interface,
+};
 use tokio::{sync::Mutex, task::AbortHandle};
 
 use crate::providers::{
   interval_provider::IntervalProvider,
-  network::{NetworkInterface, NetworkVariables},
+  network::parse_netsh::get_primary_interface_ssid_and_strength,
   variables::ProviderVariables,
 };
 
-use super::NetworkProviderConfig;
+use super::{
+  Gateway, NetworkInterface, NetworkProviderConfig, NetworkVariables,
+};
 
 pub struct NetworkProvider {
   pub config: Arc<NetworkProviderConfig>,
   abort_handle: Option<AbortHandle>,
-  networks: Arc<Mutex<Networks>>,
+  netdev_data: Arc<Mutex<NetdevData>>,
+}
+
+pub struct NetdevData {
+  pub default_interface: Arc<Mutex<Interface>>,
+  pub interfaces: Arc<Mutex<Vec<Interface>>>,
 }
 
 impl NetworkProvider {
@@ -24,7 +33,12 @@ impl NetworkProvider {
     NetworkProvider {
       config: Arc::new(config),
       abort_handle: None,
-      networks: Arc::new(Mutex::new(Networks::new_with_refreshed_list())),
+      netdev_data: Arc::new(Mutex::new(NetdevData {
+        default_interface: Arc::new(Mutex::new(
+          get_default_interface().unwrap(),
+        )),
+        interfaces: Arc::new(Mutex::new(get_interfaces())),
+      })),
     }
   }
 }
@@ -32,14 +46,14 @@ impl NetworkProvider {
 #[async_trait]
 impl IntervalProvider for NetworkProvider {
   type Config = NetworkProviderConfig;
-  type State = Mutex<Networks>;
+  type State = Mutex<NetdevData>;
 
   fn config(&self) -> Arc<NetworkProviderConfig> {
     self.config.clone()
   }
 
-  fn state(&self) -> Arc<Mutex<Networks>> {
-    self.networks.clone()
+  fn state(&self) -> Arc<Mutex<NetdevData>> {
+    self.netdev_data.clone()
   }
 
   fn abort_handle(&self) -> &Option<AbortHandle> {
@@ -52,24 +66,71 @@ impl IntervalProvider for NetworkProvider {
 
   async fn get_refreshed_variables(
     _: &NetworkProviderConfig,
-    sysinfo: &Mutex<Networks>,
+    netdev_data: &Mutex<NetdevData>,
   ) -> Result<ProviderVariables> {
-    let mut networks = sysinfo.lock().await;
-    networks.refresh();
+    let netdev = netdev_data.lock().await;
 
-    let mut interfaces = vec![];
+    let default_interface = netdev.default_interface.lock().await.clone();
 
-    for (name, data) in networks.into_iter() {
-      interfaces.push(NetworkInterface {
-        name: name.into(),
-        mac_address: data.mac_address().to_string(),
-        transmitted: data.transmitted(),
-        total_transmitted: data.total_transmitted(),
-        received: data.received(),
-        total_received: data.total_received(),
-      });
-    }
+    let interfaces = netdev.interfaces.lock().await;
 
-    Ok(ProviderVariables::Network(NetworkVariables { interfaces }))
+    #[cfg(target_os = "windows")]
+    let default_gateway_ssid_and_strength =
+      get_primary_interface_ssid_and_strength()?;
+
+    #[cfg(not(target_os = "windows"))]
+    let default_gateway_ssid_and_strength = {
+      let ssid = None;
+      let signal = None;
+      let connected = false;
+      Gateway { ssid, signal, connected };
+    };
+
+    let variables = NetworkVariables {
+      default_interface: NetworkInterface {
+        name: default_interface.name.clone(),
+        friendly_name: default_interface.friendly_name.clone(),
+        description: default_interface.description.clone(),
+        if_type: default_interface.if_type,
+        ipv4: default_interface.ipv4.clone(),
+        ipv6: default_interface.ipv6.clone(),
+        mac_addr: default_interface.mac_addr.unwrap(),
+        transmit_speed: default_interface.transmit_speed,
+        receive_speed: default_interface.receive_speed,
+        dns_servers: default_interface.dns_servers.clone(),
+        default: default_interface.default,
+      },
+      default_gateway: Gateway {
+        mac_addr: default_interface
+          .gateway
+          .as_ref()
+          .unwrap()
+          .mac_addr
+          .clone(),
+        ipv4: default_interface.gateway.as_ref().unwrap().ipv4.clone(),
+        ipv6: default_interface.gateway.as_ref().unwrap().ipv6.clone(),
+        ssid: default_gateway_ssid_and_strength.ssid.unwrap(),
+        signal_strength_percent: default_gateway_ssid_and_strength.signal.unwrap(),
+        connected: default_gateway_ssid_and_strength.connected,
+      },
+      interfaces: interfaces
+        .iter()
+        .map(|iface| NetworkInterface {
+          name: iface.name.clone(),
+          friendly_name: iface.friendly_name.clone(),
+          description: iface.description.clone(),
+          if_type: iface.if_type.clone(),
+          ipv4: iface.ipv4.clone(),
+          ipv6: iface.ipv6.clone(),
+          mac_addr: iface.mac_addr.unwrap().clone(),
+          transmit_speed: iface.transmit_speed,
+          receive_speed: iface.receive_speed,
+          dns_servers: iface.dns_servers.clone(),
+          default: iface.default,
+        })
+        .collect(),
+    };
+
+    Ok(ProviderVariables::Network(variables))
   }
 }
