--- packages/desktop/src/providers/newnetwork/mod.rs
@@ -1,8 +1,0 @@
-mod config;
-mod provider;
-mod variables;
-mod parse_netsh;
-
-pub use config::*;
-pub use provider::*;
-pub use variables::*;
